+++ app/src/main/java/woowacourse/shopping/data/datasource/remote/retrofit/OrderHistoryService.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.datasource.retrofit
+package woowacourse.shopping.data.datasource.remote.retrofit
 
 import retrofit2.Call
 import retrofit2.http.GET
