+++ app/src/main/java/woowacourse/shopping/ui/selectserver/SelectServerActivity.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.ui.selectserver
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import woowacourse.shopping.data.datasource.retrofit.RetrofitClient
+import woowacourse.shopping.data.datasource.remote.retrofit.RetrofitClient
 import woowacourse.shopping.databinding.ActivitySelectServerBinding
 import woowacourse.shopping.ui.shopping.ShoppingActivity
